+++ stable/sclmisc.h
@@ -299,7 +299,13 @@ namespace sclmisc {
 		char Marker = scllocale::DefaultMarker );
 
 	void LoadXMLAndTranslateTags(
-		const rgstry::tentry__ &FileName,
+		const fnm::rName &Filename,
+		const char *Language,
+		str::string_ &String,
+		char Marker = scllocale::DefaultMarker );
+
+	void LoadXMLAndTranslateTags(
+		const rgstry::tentry__ &Filename,
 		const sclrgstry::registry_ &Registry,
 		str::string_ &String,
 		char Marker = scllocale::DefaultMarker );