+++ tools/xdhq/NJS/test.js
@@ -70,12 +70,17 @@ function showTestButton(xdh, id) {
 	handleTestButton(xdh, true);
 }
 
+function submission(xdh, id) {
+	xdh.setContent( "Pattern", xdh.getContent( "Pattern" ).toUpperCase() );
+}
+
 function main()
 {
 	xdhq.register("Connect", connect);
 	xdhq.register("Testing", testing);
 	xdhq.register("ShowTestButton", showTestButton);
 	xdhq.register("HideTestButton", hideTestButton);
+	xdhq.register("Submission", submission);
 
 	xdhq.listen(callback, "12345");
 }