+++ tools/xdhq/prtcl.h
@@ -55,6 +55,8 @@ namespace prtcl {
 		aError_1,
 		aSetLayout_1,
 		aSetCasting_1,
+		aGetContent_1,
+		aSetContent_1,
 		a_amount,
 		a_Undefined
 	};